performance: access buffer indirectly through a ref, rather than copying
This was contributed by Mark Martinec in CPAN/RT ticket #77902.
Here is his correspondance-

...
Nevertheless I do agree that the PRINT subroutine performs some
unnecessary copying. Avoiding this could save a little bit of resources.
I did some benchmarking and I'll attach a replacement MessageParser.pm,
which accesses the buffer indirectly through a ref, avoiding copying or
accessing it through a hash. If an application passes a message in large
chunks to PRINT the speedup is very small, while an application that
does so inefficiently by calling PRINT line-by-line can see some
improvement in speed.

...

While at it... Here is a refinement of my previous attachment, this time
in a form of a patch. In addition to what the previous code did, this
one also avoids one more copying of a buffer, shaving off another
millisecond or two (out of 50 ms) on larger mail (3 MB) with larger
buffers (32 kB or 64 kB).

git-svn-id: https://svn.code.sf.net/p/dkimproxy/code/Mail-DKIM/trunk@688 f38efd27-133c-0410-a3cc-a5f95e9cf04f